Move tests to set_slice_test
